Rename: 프로젝트 패키지 구조 기능별(계층별) 분리
