+++ src/main/java/subway/service/StationService.java
@@ -1,10 +1,14 @@
-package subway;
+package subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import subway.domain.Station;
+import subway.dto.StationRequest;
+import subway.dto.StationResponse;
+import subway.repository.StationRepository;
 
 @Service
 @Transactional(readOnly = true)
